Completed task #134: -Removed center button -Changed icons (Used default material icons for this moment)
