+++ lib/l10n/app_en.arb
@@ -4,7 +4,7 @@
       "description": "Simplio name"
     },
     "noWalletsLabel": "You have no wallet",
-    "gamesLabel": "Games",
+    "gamesLabel": "My Games",
     "welcomeLabel": "Welcome to Simplio",
     "enterApplicationLabel": "I have an account",
 
@@ -14,10 +14,10 @@
     "loginBtn": "Login",
     "goToAppBtn":"Go to app",
 
-    "homeTapBarLabel": "Home",
-    "portfolioTapBarLabel": "Portfolio",
-    "gamesTapBarLabel": "Games",
-    "settingsTapBarLabel": "Settings",
+    "discoveryTapBarLabel": "Discover",
+    "gamesTapBarLabel": "My Games",
+    "inventoryTapBarLabel": "Inventory",
+    "findDappsTapBarLabel": "Find Daaps",
 
     "loginScreenTitle": "Enter Simplio.",
     "loginScreenSubTitle": "It's good to see you back.",
