+++ lib/l10n/app_ja.arb
@@ -15,9 +15,9 @@
     "goToAppBtn":"両あなた自",
 
     "homeTapBarLabel": "車両あなた自身",
-    "portfolioTapBarLabel": "車両あなた自身",
     "gamesTapBarLabel": "車両あなた自身",
-    "settingsTapBarLabel": "車両あなた自身",
+    "inventoryTapBarLabel": "車両あなた自身",
+    "findDappsTapBarLabel": "車両あなた自身",
 
     "loginScreenTitle": "車両あなた自身 車両",
     "loginScreenSubTitle": "あなたは財布を持っていません",
