+++ lib/view/screens/daaps_screen.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class DaapsScreen extends StatelessWidget {
+  const DaapsScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      body: Center(
+        child: Text('DaapsScreen screen placeholder'),
+      ),
+    );
+  }
+}
