+++ lib/view/screens/games_screen.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class GamesScreen extends StatelessWidget {
+  const GamesScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      body: Center(
+        child: Text('GamesScreen screen placeholder'),
+      ),
+    );
+  }
+}
